+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Most recent releases are shown at the top. Each release shows:
 ### changed
 - add additional available options to VectorStoreFactory (#205)
 - use query_string in Elasticsearch queries (#206)
+- `extract_noun_phrases` added to `utils` module
 
 ### fixed:
 - N/A